Changed paths: M       niiskullstrip.cpp M       vollib.c M       bwlabel.c
New preliminary version of skullstripping.


git-svn-id: svn://141.35.69.214/amap@203 3806922a-3115-4c2f-bdad-02e90158041f

+++ bwlabel.c
@@ -336,7 +336,7 @@ void fill_cluster(unsigned char *bw, int dim[3])
    for (j=0; j<ttn; j++) count[j] = 0;
    
    /* count cluster sizes */
-   for (i=0; i<n; i++) if(l[i]>0) count[l[i]]++;
+   for (i=0; i<n; i++) if((int)l[i]>0) count[(int)l[i]]++;
    
    /* find index of largest cluster */
    for (j=0; j<ttn; j++)
